+++ src/matrix_util.cpp
@@ -20,8 +20,8 @@ bool EqualMatricies(const Matrix& a, const Matrix& b) {
 }
 
 void printMatrix(Matrix& a) {
-    for (int i = 0; i < a.Rows(); ++i) {
-        for (int j = 0; j < a.Columns(); ++j) {
+    for (size_t i = 0; i < a.Rows(); ++i) {
+        for (size_t j = 0; j < a.Columns(); ++j) {
             std::cout << a(i, j) << "\t";
         }
         std::cout << "\n";
@@ -35,8 +35,8 @@ void fillMatrixRandomly(Matrix& a) {
     static constexpr int range = 10;
     static std::uniform_int_distribution<int> uid(-range, range);
 
-    for (int i = 0; i < a.Rows(); ++i) {
-        for (int j = 0; j < a.Columns(); ++j) {
+    for (size_t i = 0; i < a.Rows(); ++i) {
+        for (size_t j = 0; j < a.Columns(); ++j) {
             a(i, j) = uid(generator);
         }
     }
@@ -48,19 +48,19 @@ Matrix multiplyDecomposed(const Matrix& a, const Matrix& b, const Matrix& u, con
     Matrix c(a.Rows(), b.Columns());
     std::vector<int> m(u.Columns());
 
-    for (int i = 0; i < u.Columns(); ++i) {
+    for (size_t i = 0; i < u.Columns(); ++i) {
         int a_sum = 0;
         int b_sum = 0;
-        for (int j = 0; j < u.Rows(); ++j) {
+        for (size_t j = 0; j < u.Rows(); ++j) {
             a_sum += a(j) * u(j, i);
             b_sum += b(j) * v(j, i);
         }
         m[i] = a_sum*b_sum;
     }
 
-    for (int i = 0; i < w.Rows(); ++i) {
+    for (size_t i = 0; i < w.Rows(); ++i) {
         int m_sum = 0;
-        for (int j = 0; j < w.Columns(); ++j) {
+        for (size_t j = 0; j < w.Columns(); ++j) {
             m_sum += m[j] * w(i, j);
         }
         c(i) = m_sum;
